size_t

+++ src/binaryen-c.cpp
@@ -4256,7 +4256,7 @@ char* BinaryenModuleAllocateAndWriteStackIR(BinaryenModuleRef module,
   Colors::setEnabled(colors); // restore colors state
 
   auto str = os.str();
-  const int len = str.length() + 1;
+  const size_t len = str.length() + 1;
   char* output = (char*)malloc(len);
   std::copy_n(str.c_str(), len, output);
   return output;
